added PollService

+++ src/main/kotlin/no/hvl/dat250/app/controller/PollController.kt
@@ -1,10 +1,7 @@
 package no.hvl.dat250.app.controller
 
 import no.hvl.dat250.app.API_VERSION_1
-import no.hvl.dat250.app.dto.PollRequest
-import no.hvl.dat250.app.dto.PollResponse
-import no.hvl.dat250.app.dto.toPoll
-import no.hvl.dat250.app.dto.toResponse
+import no.hvl.dat250.app.dto.*
 import no.hvl.dat250.app.exception.MissingFieldException
 import no.hvl.dat250.app.exception.NotLoggedInException
 import no.hvl.dat250.app.exception.PollNotFoundException
@@ -12,6 +9,7 @@ import no.hvl.dat250.app.exception.PollNotOwnedByUserException
 import no.hvl.dat250.app.exception.PollNotPublicException
 import no.hvl.dat250.app.repository.PollRepository
 import no.hvl.dat250.app.service.AccountService
+import no.hvl.dat250.app.service.PollService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.web.bind.annotation.DeleteMapping
@@ -31,33 +29,25 @@ import javax.validation.Valid
 class PollController {
 
   @Autowired
-  lateinit var pollRepository: PollRepository
+  private lateinit var accountService: AccountService
 
   @Autowired
-  private lateinit var accountService: AccountService
+  private lateinit var pollService: PollService
+
 
   @ExceptionHandler(NotLoggedInException::class)
   @PostMapping("/create")
-  fun createPoll(@Valid @RequestBody pollRequest: PollRequest): PollResponse {
-    if (pollRequest.question == null || pollRequest.question == "") {
-      throw MissingFieldException("question")
-    }
-    if (pollRequest.firstAnswer == null || pollRequest.firstAnswer == "") {
-      throw MissingFieldException("firstAnswer")
-    }
-    if (pollRequest.secondAnswer == null || pollRequest.secondAnswer == "") {
-      throw MissingFieldException("secondAnswer")
-    }
-    val poll = pollRepository.saveAndFlush(pollRequest.toPoll())
+  fun createPoll(@Valid @RequestBody pollRequest: PollCreateRequest): PollResponse {
+    val poll = pollRequest.toPoll()
+    pollService.createPoll(poll)
     accountService.addPoll(poll)
     return poll.toResponse()
   }
 
   @PutMapping("/{id}")
   fun updatePoll(@PathVariable id: Long, @Valid @RequestBody pollRequest: PollRequest): PollResponse {
-    val poll = pollRepository.findByIdOrNull(id) ?: throw PollNotFoundException(id)
+    val poll = pollService.getPoll(id)
 
-    // check user owns poll
     if (accountService.isNotOwnerOf(poll)) {
       throw PollNotOwnedByUserException(id)
     }
@@ -80,13 +70,16 @@ class PollController {
     if (pollRequest.private != null) {
       poll.private = pollRequest.private
     }
-    return pollRepository.saveAndFlush(poll).toResponse()
+    pollService.updatePoll(poll)
+    return poll.toResponse()
   }
+
+
 
   @ExceptionHandler(PollNotFoundException::class, PollNotPublicException::class)
   @GetMapping
   fun getPoll(@RequestParam("id") id: Long): PollResponse {
-    val poll = pollRepository.findByIdOrNull(id) ?: throw PollNotFoundException(id)
+    val poll = pollService.getPoll(id)
     if (poll.private && accountService.isNotLoggedIn) {
       throw PollNotPublicException(id)
     }
@@ -95,9 +88,12 @@ class PollController {
 
   @DeleteMapping("/{id}")
   fun deletePoll(@PathVariable id: Long): PollResponse {
-    val poll = pollRepository.findByIdOrNull(id) ?: throw PollNotFoundException(id)
+    val poll = pollService.getPoll(id)
+    if (accountService.isNotOwnerOf(poll)){
+      throw PollNotOwnedByUserException(id)
+    }
     accountService.removePoll(poll)
-    pollRepository.delete(poll)
+    pollService.delete(poll)
     return poll.toResponse()
   }
 }
